# 8333396: Use StringBuilder internally for java.text.Format.* formatting