--- a/src/java.base/share/classes/java/text/ChoiceFormat.java
+++ b/src/java.base/share/classes/java/text/ChoiceFormat.java
@@ -742,8 +742,7 @@ public class ChoiceFormat extends NumberFormat {
                     "limits and format arrays of different length.");
         }
     }
-	
-	
+
     @Override
     boolean isInternalSubclass() {
         return true;