--- a/src/java.base/share/classes/java/text/CompactNumberFormat.java
+++ b/src/java.base/share/classes/java/text/CompactNumberFormat.java
@@ -27,6 +27,7 @@ package java.text;
 import java.io.IOException;
 import java.io.InvalidObjectException;
 import java.io.ObjectInputStream;
+import java.io.UncheckedIOException;
 import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.math.RoundingMode;
@@ -596,7 +597,20 @@ public final class CompactNumberFormat extends NumberFormat {
         return format(number, result, fieldPosition.getFieldDelegate());
     }
 
-    private StringBuffer format(double number, StringBuffer result,
+    @Override
+    <T extends Appendable & CharSequence> T formatWithGeneric(double number, T result,
+                                                   FieldPosition fieldPosition) {
+        fieldPosition.setBeginIndex(0);
+        fieldPosition.setEndIndex(0);
+        return format(number, result, fieldPosition.getFieldDelegate());
+    }
+
+    @Override
+    boolean isInternalSubclass() {
+        return true;
+    }
+
+    private <T extends Appendable & CharSequence> T format(double number, T result,
             FieldDelegate delegate) {
 
         boolean nanOrInfinity = decimalFormat.handleNaN(number, result, delegate);
@@ -686,7 +700,15 @@ public final class CompactNumberFormat extends NumberFormat {
         return format(number, result, fieldPosition.getFieldDelegate());
     }
 
-    private StringBuffer format(long number, StringBuffer result, FieldDelegate delegate) {
+    @Override
+    <T extends Appendable & CharSequence> T formatWithGeneric(long number, T result,
+                                                   FieldPosition fieldPosition) {
+        fieldPosition.setBeginIndex(0);
+        fieldPosition.setEndIndex(0);
+        return format(number, result, fieldPosition.getFieldDelegate());
+    }
+
+    private <T extends Appendable & CharSequence> T format(long number, T result, FieldDelegate delegate) {
         boolean isNegative = (number < 0);
         if (isNegative) {
             number = -number;
@@ -773,7 +795,7 @@ public final class CompactNumberFormat extends NumberFormat {
         return format(number, result, fieldPosition.getFieldDelegate());
     }
 
-    private StringBuffer format(BigDecimal number, StringBuffer result,
+    private <T extends Appendable & CharSequence> T format(BigDecimal number, T result,
             FieldDelegate delegate) {
 
         boolean isNegative = number.signum() == -1;
@@ -859,7 +881,7 @@ public final class CompactNumberFormat extends NumberFormat {
         return format(number, result, fieldPosition.getFieldDelegate(), false);
     }
 
-    private StringBuffer format(BigInteger number, StringBuffer result,
+    private <T extends Appendable & CharSequence> T format(BigInteger number, T result,
             FieldDelegate delegate, boolean formatLong) {
 
         boolean isNegative = number.signum() == -1;
@@ -936,7 +958,7 @@ public final class CompactNumberFormat extends NumberFormat {
      *                 {@code NumberFormat.Field.SIGN} and
      *                 {@code NumberFormat.Field.PREFIX} fields
      */
-    private void appendPrefix(StringBuffer result, String prefix,
+    private <T extends Appendable & CharSequence> void appendPrefix(T result, String prefix,
             FieldDelegate delegate) {
         append(result, expandAffix(prefix), delegate,
                 getFieldPositions(prefix, NumberFormat.Field.PREFIX));
@@ -952,7 +974,7 @@ public final class CompactNumberFormat extends NumberFormat {
      *                 {@code NumberFormat.Field.SIGN} and
      *                 {@code NumberFormat.Field.SUFFIX} fields
      */
-    private void appendSuffix(StringBuffer result, String suffix,
+    private <T extends Appendable & CharSequence> void appendSuffix(T result, String suffix,
             FieldDelegate delegate) {
         append(result, expandAffix(suffix), delegate,
                 getFieldPositions(suffix, NumberFormat.Field.SUFFIX));
@@ -968,11 +990,15 @@ public final class CompactNumberFormat extends NumberFormat {
      * @param positions a list of {@code FieldPosition} in the given
      *                  string
      */
-    private void append(StringBuffer result, String string,
+    private <T extends Appendable & CharSequence> void append(T result, String string,
             FieldDelegate delegate, List<FieldPosition> positions) {
         if (!string.isEmpty()) {
             int start = result.length();
-            result.append(string);
+            try {
+                result.append(string);
+            } catch (IOException ioe) {
+                throw new UncheckedIOException(ioe.getMessage(), ioe);
+            }
             for (FieldPosition fp : positions) {
                 Format.Field attribute = fp.getFieldAttribute();
                 delegate.formatted(attribute, attribute,