### Performance regression of DecimalFormat.format
From the output of perf, we can see the hottest regions contain atomic instructions.  But when run with JDK 11, there is no such problem. The reason is the removed biased locking.  
The DecimalFormat uses StringBuffer everywhere, and StringBuffer itself contains many synchronized methods.
So I added support for some new methods that accept StringBuilder which is lock-free.

### Benchmark testcase
```
@BenchmarkMode(Mode.AverageTime)
@Warmup(iterations = 5, time = 500, timeUnit = TimeUnit.MILLISECONDS)
@Measurement(iterations = 10, time = 500, timeUnit = TimeUnit.MILLISECONDS)
@State(Scope.Thread)
@OutputTimeUnit(TimeUnit.NANOSECONDS)
public class JmhDecimalFormat {

    private DecimalFormat format;

    @Setup(Level.Trial)
    public void setup() {
        format = new DecimalFormat("#0.00000");
    }

    @Benchmark
    public void testNewAndFormat() throws InterruptedException {
        new DecimalFormat("#0.00000").format(9524234.1236457);
    }

    @Benchmark
    public void testNewOnly() throws InterruptedException {
        new DecimalFormat("#0.00000");
    }

    @Benchmark
    public void testFormatOnly() throws InterruptedException {
        format.format(9524234.1236457);
    }
}
```

### Test result
#### Current JDK before optimize
```
 Benchmark                             Mode  Cnt    Score   Error  Units
JmhDecimalFormat.testFormatOnly       avgt   50  642.099 ? 1.253  ns/op
JmhDecimalFormat.testNewAndFormat     avgt   50  989.307 ? 3.676  ns/op
JmhDecimalFormat.testNewOnly          avgt   50  303.381 ? 5.252  ns/op
```


#### Current JDK after optimize
```
Benchmark                          Mode  Cnt    Score   Error  Units
JmhDecimalFormat.testFormatOnly    avgt   50  351.499 ? 0.761  ns/op
JmhDecimalFormat.testNewAndFormat  avgt   50  615.145 ? 2.478  ns/op
JmhDecimalFormat.testNewOnly       avgt   50  209.874 ? 9.951  ns/op
```

### JDK 11 
```
Benchmark                          Mode  Cnt    Score   Error  Units
JmhDecimalFormat.testFormatOnly    avgt   50  364.214 ? 1.191  ns/op
JmhDecimalFormat.testNewAndFormat  avgt   50  658.699 ? 2.311  ns/op
JmhDecimalFormat.testNewOnly       avgt   50  248.300 ? 5.158  ns/op
```

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must be properly reviewed (1 review required, with at least 1 [Reviewer](https://openjdk.org/bylaws#reviewer))
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue
- [x] Change requires CSR request [JDK-8335834](https://bugs.openjdk.org/browse/JDK-8335834) to be approved

### Issues
 * [JDK-8333396](https://bugs.openjdk.org/browse/JDK-8333396): Use StringBuilder internally for java.text.Format.* formatting (**Bug** - P3)
 * [JDK-8335834](https://bugs.openjdk.org/browse/JDK-8335834): Use StringBuilder internally for java.text.Format.* formatting (**CSR**)


### Reviewers
 * [Justin Lu](https://openjdk.org/census#jlu) (@justin-curtis-lu - Committer) 🔄 Re-review required (review applies to [5fda8747](https://git.openjdk.org/jdk/pull/19513/files/5fda874798f84ac1c1efc032e00b7c66037fe815))
 * [Naoto Sato](https://openjdk.org/census#naoto) (@naotoj - **Reviewer**)
 * [Chen Liang](https://openjdk.org/census#liach) (@liach - **Reviewer**)

### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk.git pull/19513/head:pull/19513` \
`$ git checkout pull/19513`

Update a local copy of the PR: \
`$ git checkout pull/19513` \
`$ git pull https://git.openjdk.org/jdk.git pull/19513/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 19513`

View PR using the GUI difftool: \
`$ git pr show -t 19513`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk/pull/19513.diff">https://git.openjdk.org/jdk/pull/19513.diff</a>

</details>


### Webrev
[Link to Webrev Comment](https://git.openjdk.org/jdk/pull/19513#issuecomment-2144249140)