@kuaiwei 2024-06-04
Hi, I think you can add the jmh test case.