@kuaiwei 2024-06-04
> > Hi, I think you can add the jmh test case.
> 
> [#19513 (comment)](https://github.com/openjdk/jdk/pull/19513#issue-2330131051) already contains the test case.

I mean it can be added as a test case in test/micro and be evaluated in future build.