@kuaiwei 2024-06-04
> > > > Hi, I think you can add the jmh test case.
> > > 
> > > 
> > > [#19513 (comment)](https://github.com/openjdk/jdk/pull/19513#issue-2330131051) already contains the test case.
> > 
> > 
> > I mean it can be added as a test case in test/micro and be evaluated in future build.
> 
> Thanks for your remainder. There exist a JMH test case which named 'test/micro/org/openjdk/bench/java/text/DefFormatterBench.java'.
> 
> This JMH test case run in throughput mode. If run without the patch, the output is:
> 
> ```
> Benchmark                                  Mode  Cnt     Score    Error   Units
> DefFormatterBench.testDefNumberFormatter  thrpt    5  1842.666 ? 83.694  ops/ms
> ```
> 
> If run with after apply the patch, the output is:
> 
> ```
> Benchmark                                  Mode  Cnt     Score     Error   Units
> DefFormatterBench.testDefNumberFormatter  thrpt    5  3988.274 ? 513.434  ops/ms
> ```
> 
> The score is increased by (3988.274-1842.666)/1842.666 = 116%.

Yes, it's what I expected.