@naotoj 2024-06-04
Also it would be helpful to compare the performance without biased locking in JDK11.