@bchristi-git 2024-06-04
(Same comment as on the [other PR](https://github.com/openjdk/jdk/pull/19534#issuecomment-2148568212).)

Would it be better if the benchmarks returned the created DecimalFormats ?
Just thinking about dead code elimination:
https://github.com/openjdk/jmh/blob/master/jmh-samples/src/main/java/org/openjdk/jmh/samples/JMHSample_08_DeadCode.java