@lingjun-cg 2024-06-05
> Also it would be helpful to compare the performance without biased locking in JDK11.

If run in JDK11 without biased locking, the performance is as same as run in current JDK.