@lingjun-cg 2024-06-12
@naotoj The above contains the result of JDK11 with disabled and enabled biased locking. It seems the key is the biased locking. Now our application use DecimalFormat.format() seriously,  there performance is declining when migrate from JDK 11 to JDK 21.  Could you give some suggestion?Thanks.