@lingjun-cg 2024-06-14
> I second Justin's suggestion here. The change should benefit every implementation in the JDK, not only NumberFormat. Also, I might suggest renaming the new class, as it is kind of long/redundant. How about using something as simple as "StringBuf"?

Thanks for your comment. The long name bother me for a while. I have changed it to ""StringBuf".